delete message functionality done

+++ studybud/base/views.py
@@ -90,10 +90,12 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+
 @login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
-    form = RoomForm(request.POST)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
     if (request.method == 'POST'):
         room.delete()
         return redirect('home')
@@ -149,3 +151,15 @@ def registerPage(request):
     form = UserCreationForm()
     context = {'form': form}
     return render(request, 'base/login_register.html', context)
+
+
+@login_required(login_url='login')
+def deleteMessage(request, pk):
+    message = Message.objects.get(id=pk)
+    if request.user != message.user:
+        return HttpResponse('You are not allowed here')
+    if (request.method == 'POST'):
+        message.delete()
+        return redirect('home')
+
+    return render(request, 'base/delete.html', context={'obj': message})
